Remove cli init file
